added warning if run names are identical (#79)
little torn whether to code a new name when they're identical in the `compare_runs()` function or throw the error like how this is now. ultimately i think this is better the option to encourage people to not have the same run name in their database which should be a better practice.  

closes #76